contiv-crd: Refactor validator and naming changes in controller.
  * Split validator into separate packages for l2 and l3 validation
  * Renamed object and event variables in both crd controller to avoid
    conflicts with "new" keyword
  * Handlers on object update take the old and new resource as arguments
  * Removed not used dependency `k8sClient` from crd plugin and controllers
