+++ plugins/crd/controller/telemetry/telemetry_controller.go
@@ -54,7 +54,6 @@ var serverStartTime time.Time
 type Controller struct {
 	Deps
 
-	K8sClient *kubernetes.Clientset
 	CrdClient *crdClientSet.Clientset
 	APIClient *apiextcs.Clientset
 
@@ -93,7 +92,7 @@ type Event struct {
 // Init performs the initialization of Telemetry Controller
 func (c *Controller) Init() error {
 
-	var newEvent Event
+	var event Event
 
 	c.Log.Info("Telemetry-Controller: initializing...")
 
@@ -120,29 +119,29 @@ func (c *Controller) Init() error {
 	// Add event handlers to handle the three types of events for resources (add, update, delete)
 	c.telemetryInformer.Informer().AddEventHandler(k8sCache.ResourceEventHandlerFuncs{
 		AddFunc: func(obj interface{}) {
-			newEvent.key, err = k8sCache.MetaNamespaceKeyFunc(obj)
-			newEvent.eventType = "create"
-			c.Log.Infof("Add Telemetry resource with key: %s", newEvent.key)
+			event.key, err = k8sCache.MetaNamespaceKeyFunc(obj)
+			event.eventType = "create"
+			c.Log.Infof("Add Telemetry resource with key: %s", event.key)
 			if err == nil {
-				c.queue.Add(newEvent)
+				c.queue.Add(event)
 			}
 		},
-		UpdateFunc: func(old, new interface{}) {
-			newEvent.key, err = k8sCache.MetaNamespaceKeyFunc(new)
-			newEvent.resource = new
-			newEvent.oldResource = old
-			newEvent.eventType = "update"
-			c.Log.Infof("Update Telemetry resource with key: %s", newEvent.key)
+		UpdateFunc: func(oldObj, newObj interface{}) {
+			event.key, err = k8sCache.MetaNamespaceKeyFunc(newObj)
+			event.resource = newObj
+			event.oldResource = oldObj
+			event.eventType = "update"
+			c.Log.Infof("Update Telemetry resource with key: %s", event.key)
 			if err == nil {
-				c.queue.Add(newEvent)
+				c.queue.Add(event)
 			}
 		},
 		DeleteFunc: func(obj interface{}) {
-			newEvent.key, err = k8sCache.DeletionHandlingMetaNamespaceKeyFunc(obj)
-			newEvent.eventType = "delete"
-			c.Log.Infof("Delete Telemetry resource with key: %s", newEvent.key)
+			event.key, err = k8sCache.DeletionHandlingMetaNamespaceKeyFunc(obj)
+			event.eventType = "delete"
+			c.Log.Infof("Delete Telemetry resource with key: %s", event.key)
 			if err == nil {
-				c.queue.Add(newEvent)
+				c.queue.Add(event)
 			}
 		},
 	})
@@ -197,23 +196,23 @@ func (c *Controller) runWorker() {
 func (c *Controller) processNextItem() bool {
 	// get the next item (blocking) from the queue and process or
 	// quit if shutdown requested
-	newEvent, quit := c.queue.Get()
+	event, quit := c.queue.Get()
 	if quit {
 		return false
 	}
-	defer c.queue.Done(newEvent)
+	defer c.queue.Done(event)
 
-	err := c.processItem(newEvent.(Event))
+	err := c.processItem(event.(Event))
 	if err == nil {
 		// If there is no error reset the rate limit counters
-		c.queue.Forget(newEvent)
-	} else if c.queue.NumRequeues(newEvent) < maxRetries {
-		c.Log.Errorf("Error processing %s (will retry): %v", newEvent.(Event).key, err)
-		c.queue.AddRateLimited(newEvent)
+		c.queue.Forget(event)
+	} else if c.queue.NumRequeues(event) < maxRetries {
+		c.Log.Errorf("Error processing %s (will retry): %v", event.(Event).key, err)
+		c.queue.AddRateLimited(event)
 	} else {
 		// err != nil and too many retries
-		c.Log.Errorf("Error processing %s (giving up): %v", newEvent.(Event).key, err)
-		c.queue.Forget(newEvent)
+		c.Log.Errorf("Error processing %s (giving up): %v", event.(Event).key, err)
+		c.queue.Forget(event)
 		utilruntime.HandleError(err)
 	}
 
@@ -223,35 +222,23 @@ func (c *Controller) processNextItem() bool {
 
 // processItem processes the next item from the queue and send the event update
 // to the telemetry event handler
-func (c *Controller) processItem(newEvent Event) error {
-	obj, _, err := c.telemetryInformer.Informer().GetIndexer().GetByKey(newEvent.key)
-	if err != nil {
-		return fmt.Errorf("Error fetching object with key %s from store: %v", newEvent.key, err)
-	}
-	// get object's metadata
-	objectMeta := utils.GetObjectMetaData(obj)
+func (c *Controller) processItem(event Event) error {
 
 	// process events based on its type
-	switch newEvent.eventType {
+	switch event.eventType {
 	case "create":
+		// get object's metadata
+		objectMeta := utils.GetObjectMetaData(event.resource)
 		// compare CreationTimestamp and serverStartTime and alert only on latest events
 		if objectMeta.CreationTimestamp.Sub(serverStartTime).Seconds() > 0 {
-			c.eventHandler.ObjectCreated(obj)
+			c.eventHandler.ObjectCreated(event.resource)
 			return nil
 		}
 	case "update":
-		updateEvent := Event{
-			key:         newEvent.key,
-			oldResource: newEvent.oldResource,
-			resource:    newEvent.resource,
-		}
-		c.eventHandler.ObjectUpdated(updateEvent)
+		c.eventHandler.ObjectUpdated(event.oldResource, event.resource)
 		return nil
 	case "delete":
-		deleteEvent := Event{
-			key: newEvent.key,
-		}
-		c.eventHandler.ObjectDeleted(deleteEvent)
+		c.eventHandler.ObjectDeleted(event.resource)
 		return nil
 	}
 	return nil
